requirements.txt
python-telegram-bot[webhooks]==20.7
Pillow==10.4.0